refactor: move hex-prefix helpers to hp.rs
